move it to external class to represent the tab
